+++ test/unit/rx-storage-implementations.test.ts
@@ -1646,10 +1646,10 @@ config.parallel('rx-storage-implementations.test.ts (implementation: ' + config.
                 let checkpoint: any;
                 async function getChanges(): Promise<RxDocumentData<{ key: string }>[]> {
                     const res = await storageInstance.getChangedDocumentsSince(10, checkpoint);
-                    if (res.length > 0) {
-                        checkpoint = lastOfArray(res).checkpoint;
+                    if (res.documents.length > 0) {
+                        checkpoint = res.checkpoint;
                     }
-                    return res.map(r => r.document);
+                    return res.documents;
                 }
 
                 // should not return anything if nothing has happened
@@ -1696,9 +1696,9 @@ config.parallel('rx-storage-implementations.test.ts (implementation: ' + config.
 
                 // get only the last change when requesting with empty checkpoint
                 const resTotal = await storageInstance.getChangedDocumentsSince(100);
-                assert.strictEqual(resTotal.length, 1);
-                assert.strictEqual(resTotal[0].document.key, 'foobar');
-                assert.strictEqual(resTotal[0].document._deleted, true);
+                assert.strictEqual(resTotal.documents.length, 1);
+                assert.strictEqual(resTotal.documents[0].key, 'foobar');
+                assert.strictEqual(resTotal.documents[0]._deleted, true);
 
                 storageInstance.close();
             });
@@ -1746,10 +1746,10 @@ config.parallel('rx-storage-implementations.test.ts (implementation: ' + config.
 
                 // should return both documents when called without checkpoint
                 const resultWithoutCheckpoint = await storageInstance.getChangedDocumentsSince(10);
-                assert.strictEqual(resultWithoutCheckpoint.length, 2);
+                assert.strictEqual(resultWithoutCheckpoint.documents.length, 2);
                 // the foobar-doc must have the latest value
-                const foobarRow = resultWithoutCheckpoint.find(row => row.document.key === 'foobar');
-                assert.strictEqual(ensureNotFalsy(foobarRow).document.value, '10');
+                const foobarRow = resultWithoutCheckpoint.documents.find(doc => doc.key === 'foobar');
+                assert.strictEqual(ensureNotFalsy(foobarRow).value, '10');
 
                 // insert many more documents
                 const insertManyResult = await storageInstance.bulkWrite(
@@ -1762,16 +1762,16 @@ config.parallel('rx-storage-implementations.test.ts (implementation: ' + config.
 
                 // should return both documents when called without checkpoint
                 const resultManyWithoutCheckpoint = await storageInstance.getChangedDocumentsSince(100);
-                assert.strictEqual(resultManyWithoutCheckpoint.length, 12);
+                assert.strictEqual(resultManyWithoutCheckpoint.documents.length, 12);
 
 
                 // first get 5 and then another 5 and then again.
                 const resultFirstFive = await storageInstance.getChangedDocumentsSince(5);
-                const resultSecondFive = await storageInstance.getChangedDocumentsSince(5, lastOfArray(resultFirstFive).checkpoint);
-                const resultThirdFive = await storageInstance.getChangedDocumentsSince(5, lastOfArray(resultSecondFive).checkpoint);
-                assert.strictEqual(resultFirstFive.length + resultSecondFive.length + resultThirdFive.length, 12);
-                const resultFourthFive = await storageInstance.getChangedDocumentsSince(5, lastOfArray(resultThirdFive).checkpoint);
-                assert.strictEqual(resultFourthFive.length, 0);
+                const resultSecondFive = await storageInstance.getChangedDocumentsSince(5, resultFirstFive.checkpoint);
+                const resultThirdFive = await storageInstance.getChangedDocumentsSince(5, resultSecondFive.checkpoint);
+                assert.strictEqual(resultFirstFive.documents.length + resultSecondFive.documents.length + resultThirdFive.documents.length, 12);
+                const resultFourthFive = await storageInstance.getChangedDocumentsSince(5, resultThirdFive.checkpoint);
+                assert.strictEqual(resultFourthFive.documents.length, 0);
 
 
                 // delete the document
@@ -1789,12 +1789,36 @@ config.parallel('rx-storage-implementations.test.ts (implementation: ' + config.
                 ], testContext);
                 assert.deepStrictEqual(deleteResult.error, {});
 
-                const resultAfterDelete = await storageInstance.getChangedDocumentsSince(5, lastOfArray(resultThirdFive).checkpoint);
-                assert.strictEqual(resultAfterDelete.length, 1);
-                assert.strictEqual(resultAfterDelete[0].document._deleted, true);
+                const resultAfterDelete = await storageInstance.getChangedDocumentsSince(5, resultThirdFive.checkpoint);
+                assert.strictEqual(resultAfterDelete.documents.length, 1);
+                assert.strictEqual(resultAfterDelete.documents[0]._deleted, true);
 
                 storageInstance.close();
             });
+            // TODO
+            // it('should be possible to stack up the checkpoints', async () => {
+            //     const storageInstance = await config.storage.getStorage().createStorageInstance<TestDocType>({
+            //         databaseInstanceToken: randomCouchString(10),
+            //         databaseName: randomCouchString(12),
+            //         collectionName: randomCouchString(12),
+            //         schema: getTestDataSchema(),
+            //         options: {},
+            //         multiInstance: false
+            //     });
+
+            //     const checkpoints: any[] = [];
+            //     while (checkpoints.length < 10) {
+            //         const insertResult = await storageInstance.bulkWrite([
+            //             {
+            //                 document: getWriteData({ key: 'foobar', value: '0' })
+            //             }
+            //         ], testContext);
+            //         const changesSince = await storageInstance.getChangedDocumentsSince(1, lastOfArray(checkpoints));
+            //         checkpoints.push(changesSince.checkpoint);
+            //     }
+
+            //     storageInstance.close();
+            // });
         });
         describe('.changeStream()', () => {
             it('should emit exactly one event on write', async () => {
@@ -1873,7 +1897,7 @@ config.parallel('rx-storage-implementations.test.ts (implementation: ' + config.
                     lastEvent.checkpoint
                 );
                 assert.strictEqual(
-                    emptyResult.length,
+                    emptyResult.documents.length,
                     0
                 );
 
@@ -2191,7 +2215,7 @@ config.parallel('rx-storage-implementations.test.ts (implementation: ' + config.
                 assert.ok(!(firstEventAttachment as any).data);
 
                 const changesResult = await storageInstance.getChangedDocumentsSince(1000);
-                const firstChange = changesResult[0].document;
+                const firstChange = changesResult.documents[0];
                 if (!firstChange) {
                     throw new Error('first change missing');
                 }